fix(grid): Fixing the separate imports #5701
