+++ projects/igniteui-angular/src/lib/grids/hierarchical-grid/hierarchical-grid.module.ts
@@ -2,7 +2,6 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { IgxChipsModule } from '../../chips/chips.module';
-import { IgxGridCommonModule } from '../grid-common.module';
 import { IgxGridModule } from '../grid/grid.module';
 import { IgxHierarchicalGridComponent } from './hierarchical-grid.component';
 import { IgxHierarchicalRowComponent } from './hierarchical-row.component';
@@ -13,7 +12,7 @@ import { IgxHierarchicalGridCellComponent } from './hierarchical-cell.component'
 import { IgxHierarchicalSelectionAPIService } from './selection';
 import { IgxRowIslandAPIService } from './row-island-api.service';
 import { IgxSelectModule } from '../../select/index';
-import { IgxGridComponent } from '../grid';
+import { IgxGridComponent } from '../grid/grid.component';
 
 /**
  * @hidden
